requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.35
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
slowapi==0.1.9
redis==5.0.1
pytest==7.4.3
httpx==0.25.2
prometheus-client==0.20.0
psycopg2-binary==2.9.9
pydantic-settings==2.0.3
alembic==1.13.1
bleach==6.1.0